requirements.txt
flask==2.3.3
gunicorn==21.2.0
psycopg2-binary==2.9.7
python-dotenv==1.0.0
firebase-admin==6.4.0
requests==2.31.0
pyrebase4==4.9.0
python-dateutil==2.8.2